Avoid undefined behaviour in NumericCast unit test
Because the maximum value of an int32_t isn't exactly representable
by a 32-bit float, casting it to a float produces a greater value,
and then casting that back to an int32_t is undefined. (In tested
implementations the cast saturates, so accidentally the correct
result is returned).

Using double for the intermediate value instead gives a high
enough precision that this test case can be safely expressed.

Silences warning from Undefioned Behaviour Sanitiser.

+++ src/armnn/test/UtilityTests.cpp
@@ -202,7 +202,7 @@ TEST_CASE("NumericCast")
 
     CHECK_NOTHROW(numeric_cast<int8_t>(static_cast<float>(int8_max)));
     CHECK_NOTHROW(numeric_cast<int16_t>(static_cast<float>(int16_max)));
-    CHECK_NOTHROW(numeric_cast<int32_t>(static_cast<float>(int32_max)));
+    CHECK_NOTHROW(numeric_cast<int32_t>(static_cast<double>(int32_max)));
 
     CHECK_THROWS_AS(numeric_cast<int8_t>(float_max), std::bad_cast);
     CHECK_THROWS_AS(numeric_cast<int16_t>(float_max), std::bad_cast);
